Rename report filters storage helpers file
